refactor(test): consolidate tile animation tests into particles test file
Move tile animation unit tests from separate files into test_particles.cpp to reduce file count and improve test organization. This consolidation maintains all test coverage while simplifying the project structure.

Also add environment configuration for Windows debugging by setting MinGW toolchain path in VS Code launch settings and adding a PATH fix script for native test environment.
