+++ platformio.ini
@@ -83,6 +83,7 @@ build_flags =
 
 [env:native_test]
 extends = base_native
+;extra_scripts = pre:fix_env.py
 test_framework = unity
 test_build_src = true
 test_filter = unit/*
